Group_17/requirements.txt
Required following modules/packages:
selenium
beautifulsoup4
chromedriver, and add its path in flights.py ,hotel.py,hotel_details.py in Book folder 
twilio
social-django
django-rest-framework